Organize components into a folder
