+++ client/package.json
@@ -23,7 +23,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "relay": "relay-compiler --src ./src --schema ./data/schema.graphql --extensions js jsx ts tsx --language typescript --artifactDirectory ./src/__generated__ --exclude ./src/__generated__ ./src/stories --watch",
+    "relay": "relay-compiler --src ./src/components --schema ./data/schema.graphql --extensions js jsx ts tsx --language typescript --artifactDirectory ./src/components/__generated__ --exclude ./src/components/__generated__ ./src/components/stories --watch",
     "storybook": "start-storybook -p 9009 -s public",
     "build-storybook": "build-storybook -s public"
   },
